feat: Add extra `loadImage(...)` and `createImageLoader(...)` functions
