+++ packages/react-native-nitro-image/src/useImage.ts
@@ -1,54 +1,7 @@
 import { useEffect, useState } from "react";
-import { Image as RNImage } from "react-native";
-import {
-    type AsyncImageSource,
-    isHybridImage,
-    isHybridImageLoader,
-    isHybridObject,
-} from "./AsyncImageSource";
-import { Images } from "./Images";
-import { OptionalWebImages } from "./OptionalWebLoader";
+import { type AsyncImageSource, isHybridObject } from "./AsyncImageSource";
+import { loadImage } from "./loadImage";
 import type { Image } from "./specs/Image.nitro";
-
-function createLoader(
-    source: AsyncImageSource,
-): (() => Promise<Image>) | undefined {
-    if (typeof source === "number") {
-        // It's a require(...) - a `number` which we need to resolve first
-        const resolvedSource = RNImage.resolveAssetSource(source);
-        if (__DEV__) {
-            // In debug, assets are streamed over the network
-            return createLoader({ url: resolvedSource.uri });
-        } else {
-            // In release, assets are resource IDs
-            return createLoader({ resource: resolvedSource.uri });
-        }
-    } else if (isHybridImage(source)) {
-        // don't do anything if this already is a HybridImage
-        return;
-    } else if (isHybridImageLoader(source)) {
-        // It's an ImageLoader
-        return () => source.loadImage();
-    } else if ("filePath" in source) {
-        // It's a { filePath }
-        return () => Images.loadFromFileAsync(source.filePath);
-    } else if ("arrayBuffer" in source) {
-        // It's a { arrayBuffer }
-        return () => Images.loadFromArrayBufferAsync(source.arrayBuffer);
-    } else if ("resource" in source) {
-        // It's a { resource }
-        return () => Images.loadFromResourcesAsync(source.resource);
-    } else if ("symbolName" in source) {
-        // It's a { symbolName }
-        return () => Promise.resolve(Images.loadFromSymbol(source.symbolName));
-    } else if ("url" in source) {
-        // It's a { url }
-        return () =>
-            OptionalWebImages.loadFromURLAsync(source.url, source.options);
-    } else {
-        throw new Error(`Unknown Image source! ${JSON.stringify(source)}`);
-    }
-}
 
 /**
  * A hook to asynchronously load an image from the
@@ -63,11 +16,11 @@ export function useImage(source: AsyncImageSource): Image | undefined {
 
     // biome-ignore lint: The dependencies array is a bit hacky.
     useEffect(() => {
-        const loader = createLoader(source);
-        if (loader != null) {
-            loader().then((i) => setImage(i));
-        }
+        (async () => {
+            const result = await loadImage(source);
+            setImage(result);
+        })();
     }, [isHybridObject(source) ? source : JSON.stringify(source)]);
 
-    return isHybridImage(source) ? source : image;
+    return image;
 }
