+++ packages/react-native-nitro-image/src/useImageLoader.ts
@@ -1,13 +1,6 @@
 import { useMemo } from "react";
-import { Image as RNImage } from "react-native";
-import {
-    type AsyncImageSource,
-    isHybridImage,
-    isHybridImageLoader,
-    isHybridObject,
-} from "./AsyncImageSource";
-import { ImageLoaders } from "./ImageLoaders";
-import { OptionalWebImages } from "./OptionalWebLoader";
+import { type AsyncImageSource, isHybridObject } from "./AsyncImageSource";
+import { createImageLoader } from "./createImageLoader";
 import type { Image } from "./specs/Image.nitro";
 import type { ImageLoader } from "./specs/ImageLoader.nitro";
 
@@ -15,42 +8,8 @@ export function useImageLoader(
     source: AsyncImageSource,
 ): Image | ImageLoader | undefined {
     // biome-ignore lint: The dependencies array is a bit hacky.
-    return useMemo<Image | ImageLoader | undefined>(() => {
-        if (typeof source === "number") {
-            // It's a require(...) - a `number` which we need to resolve first
-            const resolvedSource = RNImage.resolveAssetSource(source);
-            if (__DEV__) {
-                // In debug, assets are streamed over the network
-                return OptionalWebImages.createWebImageLoader(
-                    resolvedSource.uri,
-                );
-            } else {
-                // In release, assets are resource IDs
-                return ImageLoaders.createResourceImageLoader(
-                    resolvedSource.uri,
-                );
-            }
-        } else if (isHybridImage(source)) {
-            return source;
-        } else if (isHybridImageLoader(source)) {
-            return source;
-        } else if ("filePath" in source) {
-            return ImageLoaders.createFileImageLoader(source.filePath);
-        } else if ("arrayBuffer" in source) {
-            return ImageLoaders.createArrayBufferImageLoader(
-                source.arrayBuffer,
-            );
-        } else if ("resource" in source) {
-            return ImageLoaders.createResourceImageLoader(source.resource);
-        } else if ("symbolName" in source) {
-            return ImageLoaders.createSymbolImageLoader(source.symbolName);
-        } else if ("url" in source) {
-            return OptionalWebImages.createWebImageLoader(
-                source.url,
-                source.options,
-            );
-        } else {
-            return undefined;
-        }
-    }, [isHybridObject(source) ? source : JSON.stringify(source)]);
+    return useMemo<Image | ImageLoader | undefined>(
+        () => createImageLoader(source),
+        [isHybridObject(source) ? source : JSON.stringify(source)],
+    );
 }
